refactor(tv show item details page): move tv shows item details into a seperate component
move tv shows item details into a seperate component and make episode X a method
