+++ locales/en-US.json
@@ -2,6 +2,7 @@
   "badRequest": "Bad request. Try again",
   "continueListening": "Continue listening",
   "continueWatching": "Continue watching",
+  "episode": "Episode",
   "home": "Home",
   "incorrectUsernameOrPassword": "Incorrect username or password",
   "login": "Login",
